# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Client Wealth Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /clients:
    get:
      operationId: listClients
      tags: [clients]
      summary: List all clients (admin only)
      responses:
        "200":
          description: List of clients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Client"
    post:
      operationId: createClient
      tags: [clients]
      summary: Create a new client account (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClientBody"
      responses:
        "201":
          description: Client created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /clients/{clientId}:
    get:
      operationId: getClient
      tags: [clients]
      summary: Get a client by ID
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Client details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
        "404":
          description: Client not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateClient
      tags: [clients]
      summary: Update client info (admin only)
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateClientBody"
      responses:
        "200":
          description: Updated client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
    delete:
      operationId: deleteClient
      tags: [clients]
      summary: Delete a client (admin only)
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Client deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /clients/{clientId}/family-members:
    get:
      operationId: listFamilyMembers
      tags: [family]
      summary: List all family members for a client
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of family members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FamilyMember"
    post:
      operationId: createFamilyMember
      tags: [family]
      summary: Add a family member (admin only)
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFamilyMemberBody"
      responses:
        "201":
          description: Family member created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FamilyMember"
  /clients/{clientId}/family-members/{familyMemberId}:
    put:
      operationId: updateFamilyMember
      tags: [family]
      summary: Update a family member (admin only)
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: integer
        - name: familyMemberId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateFamilyMemberBody"
      responses:
        "200":
          description: Updated family member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FamilyMember"
    delete:
      operationId: deleteFamilyMember
      tags: [family]
      summary: Delete a family member (admin only)
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: integer
        - name: familyMemberId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Family member deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /clients/{clientId}/summary:
    get:
      operationId: getClientSummary
      tags: [dashboard]
      summary: Get financial summary (net worth, total assets, total liabilities, asset breakdown)
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Financial summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientSummary"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
...
    CreateClientBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        username:
          type: string
        password:
          type: string
      required:
        - name
        - username
        - password
    UpdateClientBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
    Client:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        username:
          type: string
        createdAt:
          type: string
      required:
        - id
        - name
        - username
        - createdAt
    AssetType:
      type: string
      enum:
        - mutual_fund
        - stock
        - fixed_deposit
        - recurring_deposit
        - provident_fund
        - cash_bank
...
    AssetBreakdownItem:
      type: object
      properties:
        assetType:
          $ref: "#/components/schemas/AssetType"
        total:
          type: number
        count:
          type: integer
      required:
        - assetType
        - total
        - count
    ClientSummary:
      type: object
      properties:
        clientId:
          type: integer
        clientName:
          type: string
        totalAssets:
          type: number
        totalLiabilities:
          type: number
        netWorth:
          type: number
        assetBreakdown:
          type: array
          items:
            $ref: "#/components/schemas/AssetBreakdownItem"
      required:
        - clientId
        - clientName
        - totalAssets
        - totalLiabilities
        - netWorth
        - assetBreakdown
...
    FamilyMember:
      type: object
      properties:
        id:
          type: integer
        clientId:
          type: integer
        name:
          type: string
        dob:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        relation:
          type: string
          enum: [Parent, Spouse, Child]
        createdAt:
          type: string
      required:
        - id
        - clientId
        - name
        - relation
        - createdAt
    CreateFamilyMemberBody:
      type: object
      properties:
        name:
          type: string
        dob:
          type: string
        phone:
          type: string
        relation:
          type: string
          enum: [Parent, Spouse, Child]
      required:
        - name
        - relation
    UpdateFamilyMemberBody:
      type: object
      properties:
        name:
          type: string
        dob:
          type: string
        phone:
          type: string
        relation:
          type: string
          enum: [Parent, Spouse, Child]